requirements.txt
streamlit
google-generativeai
deep-translator
requests
pytesseract
Pillow
psycopg2-binary
